Update ProfileViewModel.swift
Conform CoreDataManager protocol.
Add 'getUserImageData' method to use it in ViewController and get image data with the given entityName and attributeName.  -> Will call it on viewDidLoad to set profileView's profilePhoto to the one in core data.
Add 'saveImageData' method to save image to core data after image picking and uploading done successfully.